Bank Management System
This file is bank management system project and bank management system is based on a concept to generate and maintain daily payment transaction with the customer's account. main system a user has to pass through login system to get access, then only he/she  can use all the features of the system which included adding, Removing, Updating and viewing records, and also atm withdraw and deposit cash feature and check detail of existing customers. so all these features added in this program.

so how can we use and run this program:
firstly you need to download DEV c++ software for run this program
then write this program on your system and click on the run button
when we will run this program after that on that your system show outpur sreen and output will be show on the your screen will be print the Enter the password to login then you need to type password is pass. after that you required any keyword then you can type any keyword so example we can type a then show on your screen will be some features of bank management system then you can use.